backend/requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
flask-cors==3.0.10
gunicorn==20.1.0
openai==0.27.8
google-cloud-storage==2.9.0
google-cloud-texttospeech==2.14.1
google-cloud-speech==2.21.0
google-api-python-client==2.86.0
google-auth==2.18.1
ffmpeg-python==0.2.0
google-generativeai
google-cloud-aiplatform>=1.38.0
python-dotenv